Add garden onboarding configuration and scripts
- Introduced new JSON configuration for garden actions in `actions.json`, detailing community projects with timelines and media.
- Added a CSV template for garden onboarding in `garden-onboarding-example.csv`, outlining required information for community gardens.
- Created a new onboarding script in `garden-onboarding.js` to facilitate the onboarding process, including wallet creation and contract deployment.
- Removed the deprecated `garden-onboarding.csv` script to streamline the codebase and improve clarity in onboarding processes.
- Updated the `garden.json` file to define garden properties, including name, description, location, and associated gardeners and operators.
