+++ packages/contracts/config/actions.json
@@ -1,0 +1,31 @@
+{
+  "actions": [
+    {
+      "title": "Community Garden Cleanup",
+      "instructions": "bafkreiafya2q3nz5dbl4fvxphtrnmahl6hcjyvhzwcimgwnbh4wsy5kr7i",
+      "startTime": "2024-02-01T00:00:00Z",
+      "endTime": "2024-08-01T00:00:00Z",
+      "capitals": ["LIVING", "SOCIAL", "MATERIAL"],
+      "media": [
+        "bafkreiemwmci42u7cb23xacktk5nfspo5kfsbmflvh6sixjahjbdk2bsie",
+        "bafkreiafya2q3nz5dbl4fvxphtrnmahl6hcjyvhzwcimgwnbh4wsy5kr7i"
+      ]
+    },
+    {
+      "title": "Educational Workshop Series",
+      "instructions": "bafkreiXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX",
+      "startTime": "2024-03-01T00:00:00Z",
+      "endTime": "2024-09-01T00:00:00Z",
+      "capitals": ["INTELLECTUAL", "SOCIAL", "CULTURAL"],
+      "media": ["bafkreiYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYY"]
+    },
+    {
+      "title": "Simple Test Action",
+      "instructions": "bafkreitest1",
+      "startTime": "2024-01-01T00:00:00Z",
+      "endTime": "2024-12-31T23:59:59Z",
+      "capitals": ["LIVING", "SOCIAL"],
+      "media": ["bafkreimedia1"]
+    }
+  ]
+}
