+++ packages/contracts/config/garden-onboarding-example.csv
@@ -1,0 +1,11 @@
+Instructions: Fill in the garden information and participant details below
+Name,Community Learning Garden
+Description,A vibrant space where community members collaborate on environmental and educational projects
+Location,"San Francisco, CA"
+Banner Image,https://example.com/banner.jpg
+,
+Garden Operators,Gardeners
+operator1@example.com,gardener1@example.com
+operator2@example.com,gardener2@example.com
+,gardener3@example.com
+,gardener4@example.com
